feat: before new grid is generated, the existing grid is removed
I was having issues with this because I would like user to see an existing 16x16 grid already and can create a new grid if they'd like. this is a better ux. i've isolated out the adding event listener on squares so they change colours on hover to a function. this is called at the end of creating new grid rather than in main previously, as then creating a new grid would not have that event listener on it. I have to fetch the squares from DOM again rather than adding the event listener in the first while loop, because using .clone on the row node does not copy over the event listener