feat: ensure same size of total grid despite number of squares
implement flex: 1 (flex-grow: 1, flex-shrink: 1, flex-basis: 0) and width of each row being the same as height of container (to ensure it is a square). Usage of vh unit to keep relatively responsive. flex-grow: 1 is the most important as it means each square grows to fill the space, evenly. flex-shrink: 1 is important for when the squares need to be compressed. There is no width declaration, so flex-basis = 0% to ensure grow/shrink evenly. flex-basis: auto is not required